+++ lib/src/ats/ats.g.dart
@@ -13,7 +13,7 @@ _$AtsReceptionProductImpl _$$AtsReceptionProductImplFromJson(
       assetId: json['assetId'] as String?,
       fuelType: const AtsFuelTypeOrNullConverter()
           .fromJson(json['fuelType'] as String?),
-      fuelSubType: json['fuelSubType'] as String?,
+      fuelSubtype: json['fuelSubtype'] as String?,
       density: (json['density'] as num?)?.toDouble(),
       temperature: (json['temperature'] as num?)?.toDouble(),
       volumeBought: (json['volumeBought'] as num?)?.toDouble(),
@@ -32,7 +32,7 @@ Map<String, dynamic> _$$AtsReceptionProductImplToJson(
       'id': instance.id,
       'assetId': instance.assetId,
       'fuelType': const AtsFuelTypeOrNullConverter().toJson(instance.fuelType),
-      'fuelSubType': instance.fuelSubType,
+      'fuelSubtype': instance.fuelSubtype,
       'density': instance.density,
       'temperature': instance.temperature,
       'volumeBought': instance.volumeBought,